Fix bug in memory allocator causing it to fail on systems where size_t is 32 bits.
Bug: 192824793
Test: Manual

+++ test/memoryadvice/memorytest/src/main/cpp/allocator.h
@@ -15,7 +15,7 @@ class Allocator {
  public:
   Allocator() = default;
   ~Allocator() = default;
-  bool Allocate(size_t to_allocate);
+  bool Allocate(uint64_t to_allocate);
   void Release();
   void ReleasePartial(size_t bytes);
 };
